worker-service/requirements.txt
pyjwt==2.8.0
bcrypt==4.1.2
httpx==0.26.0
pydantic==2.5.3
python-dateutil==2.8.2



